Refactored some test utility functions into its own module
